better transitions; can hide the primary inputs
